pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "jbfoundry"
version = "0.1.0"
description = "An integrated framework for automatic LLM jailbreak techniques"
authors = ["JBFoundry Team"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"

# LLM Core - Always needed for API-based attacks
litellm = "^1.72.0"
openai = "^1.3.0"

# Data & Datasets - Core functionality
datasets = "^2.14.0"
huggingface-hub = "^0.19.0"
numpy = "^1.24.0"
pandas = "^2.0.0"

# NLP Utilities - Core for prompt processing
nltk = "^3.8.1"
regex = "^2023.6.3"

# CLI & Utilities - Core framework
click = "^8.1.3"
pyyaml = "^6.0"
tqdm = "^4.66.0"
colorama = "^0.4.6"
requests = "^2.32.0"

# ML Core
torch = "^2.0.0"
transformers = "^4.36.0"

# Optional: Heavy dependencies for specific attack types
peft = {version = "^0.5.0", optional = true}
accelerate = {version = "^0.32.0", optional = true}
nanogcg = {version = "^0.1.0", optional = true}

# Optional: Analysis & visualization
matplotlib = {version = "^3.7.0", optional = true}
seaborn = {version = "^0.13.0", optional = true}

# Optional: Paper preprocessing for agent workflows
PyPDF2 = {version = "^3.0.0", optional = true}
google-auth = {version = "^2.0.0", optional = true}

# Optional: Inspect AI integration for result viewing
inspect-ai = {version = "^0.3.0", optional = true}

[tool.poetry.extras]
# Fine-tuning and optimization-based attacks (GCG, AutoDAN, BEAST, etc.)
training = ["peft", "accelerate", "nanogcg"]

# Analysis tools (heatmaps, comparisons, statistics)
analysis = ["matplotlib", "seaborn"]

# Agent workflows (paper-to-attack conversion)
agents = ["PyPDF2", "google-auth"]

# Inspect AI result viewer integration
inspect = ["inspect-ai"]

# Everything (for agent deployment where all attacks must work)
all = ["peft", "accelerate", "nanogcg", "matplotlib", "seaborn", "PyPDF2", "google-auth", "inspect-ai"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.1.0"
black = "^23.3.0"
isort = "^5.12.0"
ruff = "^0.0.278"
pre-commit = "^3.3.3"

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=src/jbfoundry --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (black handles this)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # imported but unused
